+++ app/Basket/Synchronisation/MapApplicationHelper.php
@@ -65,6 +65,7 @@ class MapApplicationHelper
             $application->ext_order_amount = $orderEntity->getAmount();
             $application->ext_order_description = $orderEntity->getDescription();
             $application->ext_order_validity = $orderEntity->getValidity();
+            $application->ext_order_hold = $orderEntity->getHold();
         }
     }
 
@@ -92,6 +93,7 @@ class MapApplicationHelper
         if ($fulfilmentEntity !== null) {
             $application->ext_fulfilment_method = $fulfilmentEntity->getMethod();
             $application->ext_fulfilment_location = $fulfilmentEntity->getLocation();
+            $application->ext_fulfilment_reference = $fulfilmentEntity->getReference();
         }
     }