+++ app/Http/routes.php
@@ -163,6 +163,7 @@ Route::group(['middleware' => 'auth'], function () {
         Route::get('ajax/installations/{installation}/products/{product}/credit-info', 'AjaxController@getCreditInformationForProduct');
         Route::post('ajax/locations/{location}/profile/personal', 'AjaxController@setProfilePersonal');
         Route::post('ajax/locations/{location}/profile/address', 'AjaxController@addProfileAddress');
+        Route::post('ajax/locations/{location}/profile/removeAddress', 'AjaxController@removeProfileAddress');
         Route::post('ajax/locations/{location}/profile/employment', 'AjaxController@setProfileEmployment');
         Route::post('ajax/locations/{location}/profile/financial', 'AjaxController@setProfileFinancial');
     });