+++ database/migrations/2016_12_14_154000_add_fulfilment_reference_to_application.php
@@ -1,0 +1,36 @@
+<?php
+
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
+
+/**
+ * Add Fulfilment Reference To Application
+ *
+ * @author EB
+ */
+class AddFulfilmentReferenceToApplication extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::table('applications', function (Blueprint $table) {
+            $table->string('ext_fulfilment_reference')->nullable()->after('ext_fulfilment_location');
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::table('applications', function (Blueprint $table) {
+            $table->dropColumn('ext_fulfilment_reference');
+        });
+    }
+}