+++ database/migrations/2016_12_16_163000_add_order_hold_to_application.php
@@ -1,0 +1,36 @@
+<?php
+
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
+
+/**
+ * Add Order Hold To Application
+ *
+ * @author EB
+ */
+class AddOrderHoldToApplication extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::table('applications', function (Blueprint $table) {
+            $table->dateTimeTz('ext_order_hold')->nullable()->after('ext_order_amount');
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::table('applications', function (Blueprint $table) {
+            $table->dropColumn('ext_order_hold');
+        });
+    }
+}